+++ src/routes/DrawerRoute.tsx
@@ -2,7 +2,7 @@ import { createDrawerNavigator } from "@react-navigation/drawer";
 import React from "react";
 import { DrawerContent } from "../components";
 import { pages as p } from "../constants";
-import { Auth, HomeScreen, ProfileScreen, RoomListScreen } from "../screens";
+import { Auth, Home, Profile, RoomList } from "../screens";
 
 const Drawer = createDrawerNavigator();
 
@@ -10,14 +10,14 @@ const DrawerRoute = () => {
   return (
     <Drawer.Navigator
       drawerContent={(props) => <DrawerContent {...props} />}
-      initialRouteName={p.HomeScreen}
+      initialRouteName={p.Home}
     >
-      <Drawer.Screen name={p.HomeScreen} component={HomeScreen} />
-      <Drawer.Screen name={p.RoomListScreen} component={RoomListScreen} />
+      <Drawer.Screen name={p.Home} component={Home} />
+      <Drawer.Screen name={p.RoomList} component={RoomList} />
       <Drawer.Screen name={p.Auth} component={Auth} />
       <Drawer.Screen
-        name={p.ProfileScreen}
-        component={ProfileScreen}
+        name={p.Profile}
+        component={Profile}
         options={{ unmountOnBlur: true }}
       />
     </Drawer.Navigator>
