+++ src/screens/VideoCall/index.tsx
@@ -1,18 +1,18 @@
 import { CompositeNavigationProp } from "@react-navigation/core";
 import React, { FC, useRef, useState } from "react";
-import { Image, View, StyleSheet, ViewStyle } from "react-native";
-import { ArrowBack, Plus, Swap } from "../../assets";
-import { AppCanvas, Button, ButtonHeader } from "../components";
-import { widthPercent } from "../config";
-import { colorsPalette as cp, spacing as sp } from "../constants";
+import { Image, View } from "react-native";
 import { RNCamera } from "react-native-camera";
 import { useSelector } from "react-redux";
-import AppState from "../redux";
+import { ArrowBack, Plus, Swap } from "../../../assets";
+import { AppCanvas, Button, ButtonHeader } from "../../components";
+import { colorsPalette as cp, spacing as sp } from "../../constants";
+import AppState from "../../redux";
+import styles from "./styles";
 
-interface VideoCallScreenProps {
+interface VideoCallProps {
   navigation: CompositeNavigationProp<any, any>;
 }
-const VideoCallScreen: FC<VideoCallScreenProps> = ({ navigation }) => {
+const VideoCall: FC<VideoCallProps> = ({ navigation }) => {
   const {
     sessionReducer: { photoURL },
   } = useSelector((state: AppState) => state);
@@ -71,49 +71,4 @@ const VideoCallScreen: FC<VideoCallScreenProps> = ({ navigation }) => {
   );
 };
 
-export default VideoCallScreen;
-
-const styles = () => {
-  const base = 54;
-  const baseButton: ViewStyle = {
-    width: base,
-    height: base,
-    borderRadius: base,
-    justifyContent: "center",
-    alignItems: "center",
-  };
-  return StyleSheet.create({
-    imageProfile: { width: "100%", height: "100%" },
-    ownVideoCont: {
-      position: "absolute",
-      width: widthPercent(30),
-      height: (widthPercent(30) * 4) / 3,
-      backgroundColor: "red",
-      right: sp.sm,
-      top: sp.sm,
-    },
-    swapButton: { ...baseButton, backgroundColor: "#0005" },
-    closeButton: { ...baseButton, backgroundColor: cp.red1 },
-    buttonsCont: {
-      height: base,
-      position: "absolute",
-      justifyContent: "center",
-      alignItems: "center",
-      bottom: sp.l,
-      width: widthPercent(100),
-      flexDirection: "row",
-    },
-    backCont: {
-      backgroundColor: "#0005",
-      top: sp.sm,
-      left: sp.sm,
-      position: "absolute",
-      borderRadius: 8,
-    },
-    cameraCont: {
-      flex: 1,
-      justifyContent: "flex-end",
-      alignItems: "center",
-    },
-  });
-};
+export default VideoCall;
